PLEASE READ: Large refactor; data is now passed around in classes
Because this code takes a set of Tropomi files, we have a class
to hold data for these. As these files are aggregated into a
single product, we also have a class to hold that.

The Tropomi product, TropomiData, is constructed from the
filepaths to a matched DLR and FRESCO file. On creation, it
reads the contents of each file into arrays, filters them and
checks them for parity. It can also return the number of observations
on request.

The aggregate product, CloudVariableStore, contains methods that take a
TropomiData object and aggregate their contents in lat_lon
defined grids. There are also functions for saving the aggregate products
to netCDF and plotting.

This was done because it is easy to lose track of large colletions of
arrays - each piece of data is now stored in the relevent class. It would
be fairly straightforward to alter the filtering functions for TropomiData,
for example, or adapt to a change of format in S5P files.